Split up the different AI.
Made the different strategies different classes. AIMASTER contains the
basics and is a base class for the other classes.
